+++ services/pdfgenerator/data/index.js
@@ -108,6 +108,11 @@ export async function getRentsData(params) {
         return acc;
       }, [])
     },
+    guaranty: dbTenant.guaranty || 0,
+    guarantyPayback: dbTenant.guarantyPayback || 0,
+    guarantyType: dbTenant.guarantyType || '',
+    guarantyDate: dbTenant.guarantyDate || null,
+    guarantyReference: dbTenant.guarantyReference || '',
     rents
   };
   if (dbTenant.terminationDate) {